Check a Vulkan view's image usage against the flags Vulkan names
- Whether a texture gets a native image view is now decided from the Vulkan
  image usage it was created with, against the list the valid-usage rule
  names (sampled, storage, color, depth-stencil, input and transient
  attachment), rather than from runtime usage terms. A runtime usage added
  later that maps to one of those flags cannot fall outside the check.
- A render-pass attachment whose view has no native image view is refused
  with a message saying so, separately from one that does not resolve at all.
- The view suite's trace names each usage it tries through the usage's
  stream operator.

+++ donner/gpu/vulkan/tests/VulkanTextureView_tests.cc
@@ -67,7 +67,7 @@ TEST_F(VulkanTextureViewTest, AViewOfATextureThatCanBeViewedHasANativeView) {
   for (const TextureUsage usage :
        {TextureUsage::Sampled, TextureUsage::StorageBinding, TextureUsage::RenderAttachment,
         TextureUsage::Sampled | TextureUsage::CopySrc}) {
-    SCOPED_TRACE(testing::Message() << "usage " << static_cast<uint32_t>(usage));
+    SCOPED_TRACE(testing::Message() << "usage " << usage);
     const Texture texture = makeTexture(usage);
     const TextureView view =
         GetResultOrFail(device_->createTextureView(texture, TextureViewDescriptor{"viewable"}));
